Add support for new assets format
