--- frcData/Field2d_2018.json
@@ -1,7 +1,0 @@
-{
-  "sourceUrl": "https://www.chiefdelphi.com/t/pic-top-down-orthographic-view-of-the-field/161896/21",
-  "topLeft": [519, 78],
-  "bottomRight": [3320, 1480],
-  "widthInches": 648,
-  "heightInches": 324
-}
