--- frcData/Field2d_2019.json
@@ -1,7 +1,0 @@
-{
-  "sourceUrl": "https://www.chiefdelphi.com/t/8k-2019-top-down-orthographic-field-views/337019",
-  "topLeft": [496, 78],
-  "bottomRight": [3204, 1421],
-  "widthInches": 648,
-  "heightInches": 324
-}
