--- frcData/Field2d_2020.json
@@ -1,7 +1,0 @@
-{
-  "sourceUrl": "https://www.chiefdelphi.com/t/2020-top-down-orthographic-field-views/369892",
-  "topLeft": [519, 73],
-  "bottomRight": [3028, 1363],
-  "widthInches": 629.25,
-  "heightInches": 323.25
-}
