--- frcData/Field2d_2022.json
@@ -1,7 +1,0 @@
-{
-  "sourceUrl": "https://www.chiefdelphi.com/t/2022-top-down-field-renders/399031",
-  "topLeft": [884, 148],
-  "bottomRight": [6160, 2786],
-  "widthInches": 648,
-  "heightInches": 324
-}
