--- frcData/Field2d_2023.json
@@ -1,7 +1,0 @@
-{
-  "sourceUrl": "https://www.chiefdelphi.com/t/2023-top-down-field-renders/421365",
-  "topLeft": [482, 89],
-  "bottomRight": [3126, 1373],
-  "widthInches": 651.25,
-  "heightInches": 315.5
-}
