--- frcData/Field3d_2021.json
@@ -1,15 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/64c9a36aa14dd9954d373667/w/923414dccbd21a86b90509b3/e/ab8019f2342c82f78d7c8004",
-  "rotations": [
-    {
-      "axis": "x",
-      "degrees": 90
-    },
-    {
-      "axis": "z",
-      "degrees": 90
-    }
-  ],
-  "widthInches": 629.25,
-  "heightInches": 323.25
-}
