--- frcData/Field3d_2022.json
@@ -1,6 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/e491d0b03d0534894c2f50ee/w/746eeb7ee6910a6cbe50eb65/e/6f25d3aee97d55ef18af9901",
-  "rotations": [{ "axis": "x", "degrees": 90 }],
-  "widthInches": 648,
-  "heightInches": 324
-}
