--- frcData/Field3d_2023.json
@@ -1,9 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/418c44a447140f83b050d1e1/w/5958cbddccccb793d441b4bf/e/e84e39c6eb04228bd804d14e",
-  "rotations": [
-    { "axis": "x", "degrees": 90 },
-    { "axis": "z", "degrees": 180 }
-  ],
-  "widthInches": 651.25,
-  "heightInches": 315.5
-}
