--- frcData/Field3d_Evergreen.json
@@ -1,6 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/eeba08a27f462f02e601684e/w/0f61d86476db9c2997f2f9e2/e/d3543ebb4bf93b5d352a38fe",
-  "rotations": [{ "axis": "x", "degrees": 90 }],
-  "widthInches": 648,
-  "heightInches": 324
-}
