--- frcData/Joystick_Logitech F310.json
@@ -1,140 +1,0 @@
-[
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [1193, 541],
-    "sizePx": [95, 95],
-    "sourceIndex": 1
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [1294, 440],
-    "sizePx": [95, 95],
-    "sourceIndex": 2
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [1093, 440],
-    "sizePx": [95, 95],
-    "sourceIndex": 3
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [1196, 340],
-    "sizePx": [95, 95],
-    "sourceIndex": 4
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [410, 125],
-    "sizePx": [165, 76],
-    "sourceIndex": 5
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [1215, 125],
-    "sizePx": [165, 76],
-    "sourceIndex": 6
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [667, 338],
-    "sizePx": [85, 64],
-    "sourceIndex": 7
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [945, 340],
-    "sizePx": [85, 64],
-    "sourceIndex": 8
-  },
-  {
-    "type": "axis",
-    "isYellow": false,
-    "centerPx": [120, 204],
-    "sizePx": [80, 234],
-    "sourceIndex": 2,
-    "sourceRange": [0, 1]
-  },
-  {
-    "type": "axis",
-    "isYellow": false,
-    "centerPx": [1501, 204],
-    "sizePx": [80, 234],
-    "sourceIndex": 3,
-    "sourceRange": [0, 1]
-  },
-  {
-    "type": "joystick",
-    "isYellow": true,
-    "centerPx": [607, 663],
-    "radiusPx": 100,
-    "xSourceIndex": 0,
-    "xSourceInverted": false,
-    "ySourceIndex": 1,
-    "ySourceInverted": true,
-    "buttonSourceIndex": 9
-  },
-  {
-    "type": "joystick",
-    "isYellow": true,
-    "centerPx": [1000, 663],
-    "radiusPx": 100,
-    "xSourceIndex": 4,
-    "xSourceInverted": false,
-    "ySourceIndex": 5,
-    "ySourceInverted": true,
-    "buttonSourceIndex": 10
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [415, 365],
-    "sizePx": [70, 70],
-    "sourceIndex": 0,
-    "sourcePov": "up"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [485, 435],
-    "sizePx": [70, 70],
-    "sourceIndex": 0,
-    "sourcePov": "right"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [415, 505],
-    "sizePx": [70, 70],
-    "sourceIndex": 0,
-    "sourcePov": "down"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [345, 435],
-    "sizePx": [70, 70],
-    "sourceIndex": 0,
-    "sourcePov": "left"
-  }
-]
