--- frcData/Joystick_PS4 Controller.json
@@ -1,172 +1,0 @@
-[
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [1990, 615],
-    "sizePx": [185, 185],
-    "sourceIndex": 1
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [2165, 793],
-    "sizePx": [185, 185],
-    "sourceIndex": 2
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [2338, 616],
-    "sizePx": [185, 185],
-    "sourceIndex": 3
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [2162, 441],
-    "sizePx": [185, 185],
-    "sourceIndex": 4
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [697, 210],
-    "sizePx": [355, 100],
-    "sourceIndex": 5
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [2145, 210],
-    "sizePx": [355, 100],
-    "sourceIndex": 6
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [697, 110],
-    "sizePx": [355, 100],
-    "sourceIndex": 7
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [2145, 110],
-    "sizePx": [355, 100],
-    "sourceIndex": 8
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [935, 380],
-    "sizePx": [100, 160],
-    "sourceIndex": 9
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [1908, 380],
-    "sizePx": [100, 160],
-    "sourceIndex": 10
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [1422, 950],
-    "sizePx": [150, 150],
-    "sourceIndex": 13
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [1420, 490],
-    "sizePx": [820, 415],
-    "sourceIndex": 14
-  },
-  {
-    "type": "axis",
-    "isYellow": false,
-    "centerPx": [260, 260],
-    "sizePx": [140, 400],
-    "sourceIndex": 3,
-    "sourceRange": [0, 1]
-  },
-  {
-    "type": "axis",
-    "isYellow": false,
-    "centerPx": [2577, 260],
-    "sizePx": [140, 400],
-    "sourceIndex": 4,
-    "sourceRange": [0, 1]
-  },
-  {
-    "type": "joystick",
-    "isYellow": true,
-    "centerPx": [1040, 945],
-    "radiusPx": 150,
-    "xSourceIndex": 0,
-    "xSourceInverted": false,
-    "ySourceIndex": 1,
-    "ySourceInverted": true,
-    "buttonSourceIndex": 11
-  },
-  {
-    "type": "joystick",
-    "isYellow": true,
-    "centerPx": [1802, 945],
-    "radiusPx": 150,
-    "xSourceIndex": 2,
-    "xSourceInverted": false,
-    "ySourceIndex": 5,
-    "ySourceInverted": true,
-    "buttonSourceIndex": 12
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [680, 475],
-    "sizePx": [140, 140],
-    "sourceIndex": 0,
-    "sourcePov": "up"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [820, 615],
-    "sizePx": [140, 140],
-    "sourceIndex": 0,
-    "sourcePov": "right"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [680, 755],
-    "sizePx": [140, 140],
-    "sourceIndex": 0,
-    "sourcePov": "down"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [540, 615],
-    "sizePx": [140, 140],
-    "sourceIndex": 0,
-    "sourcePov": "left"
-  }
-]
