--- frcData/Joystick_Xbox Controller (Blue).json
@@ -1,140 +1,0 @@
-[
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [828, 329],
-    "sizePx": [80, 80],
-    "sourceIndex": 1
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [891, 267],
-    "sizePx": [80, 80],
-    "sourceIndex": 2
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [762, 263],
-    "sizePx": [80, 80],
-    "sourceIndex": 3
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [826, 201],
-    "sizePx": [80, 80],
-    "sourceIndex": 4
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [335, 100],
-    "sizePx": [208, 60],
-    "sourceIndex": 5
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [820, 100],
-    "sizePx": [208, 60],
-    "sourceIndex": 6
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [507, 262],
-    "sizePx": [60, 60],
-    "sourceIndex": 7
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": true,
-    "centerPx": [649, 262],
-    "sizePx": [60, 60],
-    "sourceIndex": 8
-  },
-  {
-    "type": "axis",
-    "isYellow": false,
-    "centerPx": [145, 150],
-    "sizePx": [60, 160],
-    "sourceIndex": 2,
-    "sourceRange": [0, 1]
-  },
-  {
-    "type": "axis",
-    "isYellow": false,
-    "centerPx": [1013, 150],
-    "sizePx": [60, 160],
-    "sourceIndex": 3,
-    "sourceRange": [0, 1]
-  },
-  {
-    "type": "joystick",
-    "isYellow": true,
-    "centerPx": [330, 264],
-    "radiusPx": 60,
-    "xSourceIndex": 0,
-    "xSourceInverted": false,
-    "ySourceIndex": 1,
-    "ySourceInverted": true,
-    "buttonSourceIndex": 9
-  },
-  {
-    "type": "joystick",
-    "isYellow": true,
-    "centerPx": [705, 409],
-    "radiusPx": 60,
-    "xSourceIndex": 4,
-    "xSourceInverted": false,
-    "ySourceIndex": 5,
-    "ySourceInverted": true,
-    "buttonSourceIndex": 10
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [451, 369],
-    "sizePx": [52, 52],
-    "sourceIndex": 0,
-    "sourcePov": "up"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [503, 421],
-    "sizePx": [52, 52],
-    "sourceIndex": 0,
-    "sourcePov": "right"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [451, 473],
-    "sizePx": [52, 52],
-    "sourceIndex": 0,
-    "sourcePov": "down"
-  },
-  {
-    "type": "button",
-    "isYellow": true,
-    "isEllipse": false,
-    "centerPx": [399, 421],
-    "sizePx": [52, 52],
-    "sourceIndex": 0,
-    "sourcePov": "left"
-  }
-]
