--- frcData/Robot_Bot-Bot Strikes Back (6328).json
@@ -1,25 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/cc566eb0195b26b436889a23/w/965bcb1860503c4949111a6d/e/bb0adeaa7ed3692762b5861a",
-  "rotations": [
-    { "axis": "x", "degrees": 90 },
-    { "axis": "z", "degrees": 90 }
-  ],
-  "position": [0.05, 0, 0.06],
-  "cameras": [
-    {
-      "name": "Limelight",
-      "rotations": [{ "axis": "y", "degrees": -30 }],
-      "position": [-0.075, 0, 0.465],
-      "resolution": [960, 720],
-      "fov": 59.6
-    },
-    {
-      "name": "Driver Camera",
-      "rotations": [{ "axis": "y", "degrees": 30 }],
-      "position": [-0.05, -0.084, 0.44],
-      "resolution": [320, 240],
-      "fov": 120
-    }
-  ],
-  "components": []
-}
