--- frcData/Robot_Duck Bot.json
@@ -1,68 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/b83cb8e60cea7c17cbee6692/w/13284a7a33c1b2621631c2ee/e/ee2cff08193b9fc024e2c99f",
-  "rotations": [{ "axis": "z", "degrees": 90 }],
-  "position": [0, 0, 0],
-  "cameras": [
-    {
-      "name": "Front Camera",
-      "rotations": [
-        {
-          "axis": "y",
-          "degrees": 20
-        }
-      ],
-      "position": [0.2, 0, 0.8],
-      "resolution": [960, 720],
-      "fov": 90
-    },
-    {
-      "name": "Back Camera",
-      "rotations": [
-        {
-          "axis": "y",
-          "degrees": 20
-        },
-        {
-          "axis": "z",
-          "degrees": 180
-        }
-      ],
-      "position": [-0.2, 0, 0.8],
-      "resolution": [960, 720],
-      "fov": 90
-    },
-    {
-      "name": "Left Camera",
-      "rotations": [
-        {
-          "axis": "y",
-          "degrees": 20
-        },
-        {
-          "axis": "z",
-          "degrees": 90
-        }
-      ],
-      "position": [0, 0.2, 0.8],
-      "resolution": [960, 720],
-      "fov": 90
-    },
-    {
-      "name": "Right Camera",
-      "rotations": [
-        {
-          "axis": "y",
-          "degrees": 20
-        },
-        {
-          "axis": "z",
-          "degrees": -90
-        }
-      ],
-      "position": [0, -0.2, 0.8],
-      "resolution": [960, 720],
-      "fov": 90
-    }
-  ],
-  "components": []
-}
