--- frcData/Robot_KitBot.json
@@ -1,57 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/eeba08a27f462f02e601684e/w/0f61d86476db9c2997f2f9e2/e/d3543ebb4bf93b5d352a38fe",
-  "rotations": [{ "axis": "z", "degrees": 90 }],
-  "position": [0.12, 3.15, 0],
-  "cameras": [
-    {
-      "name": "Front Camera",
-      "rotations": [{ "axis": "y", "degrees": 20 }],
-      "position": [0.2, 0, 0.8],
-      "resolution": [960, 720],
-      "fov": 100
-    },
-    {
-      "name": "Back Camera",
-      "rotations": [
-        { "axis": "y", "degrees": 20 },
-        { "axis": "z", "degrees": 180 }
-      ],
-      "position": [-0.2, 0, 0.8],
-      "resolution": [960, 720],
-      "fov": 100
-    },
-    {
-      "name": "Left Camera",
-      "rotations": [
-        {
-          "axis": "y",
-          "degrees": 20
-        },
-        {
-          "axis": "z",
-          "degrees": 90
-        }
-      ],
-      "position": [0, 0.2, 0.8],
-      "resolution": [960, 720],
-      "fov": 90
-    },
-    {
-      "name": "Right Camera",
-      "rotations": [
-        {
-          "axis": "y",
-          "degrees": 20
-        },
-        {
-          "axis": "z",
-          "degrees": -90
-        }
-      ],
-      "position": [0, -0.2, 0.8],
-      "resolution": [960, 720],
-      "fov": 90
-    }
-  ],
-  "components": []
-}
