--- frcData/Robot_REV 2023 Starter Bot.json
@@ -1,47 +1,0 @@
-{
-  "sourceUrl": "https://cad.onshape.com/documents/56a12b92fd203ac55406d12b/w/84a419a826bbc981204d4309/e/3c4370600ccd32582effc7bc",
-  "rotations": [
-    {
-      "axis": "x",
-      "degrees": 90
-    },
-    {
-      "axis": "z",
-      "degrees": 90
-    }
-  ],
-  "position": [0, 0, 0],
-  "cameras": [],
-  "components": [
-    {
-      "zeroedRotations": [
-        {
-          "axis": "x",
-          "degrees": 90
-        },
-        {
-          "axis": "z",
-          "degrees": 90
-        },
-        {
-          "axis": "y",
-          "degrees": 211
-        }
-      ],
-      "zeroedPosition": [0.645, 0, 0.48]
-    },
-    {
-      "zeroedRotations": [
-        {
-          "axis": "x",
-          "degrees": 90
-        },
-        {
-          "axis": "z",
-          "degrees": 90
-        }
-      ],
-      "zeroedPosition": [0.305, 0, -0.37]
-    }
-  ]
-}
