+++ package.json
@@ -65,7 +65,6 @@
       "bundles/*",
       "icons/**/*",
       "!icons/**/*.iconset",
-      "frcData/**/*",
       "docs/**/*",
       "!**/.*"
     ],
