+++ src/shared/visualizers/OdometryVisualizer.ts
@@ -48,7 +48,7 @@ export default class OdometryVisualizer implements Visualizer {
     }
 
     // Get game data and update image element
-    let gameData = window.frcData?.field2ds.find((game) => game.title == command.options.game);
+    let gameData = window.assets?.field2ds.find((game) => game.name == command.options.game);
     if (!gameData) return null;
     if (gameData.path != this.lastImageSource) {
       this.lastImageSource = gameData.path;
